pyproject.toml
[tool.poetry]
name = "gzr-zebrands-backend-test"
version = "0.1.0"
description = ""
authors = ["Gerardo Zumaya"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
uvicorn = "^0.17.4"
alembic = "^1.7"
sqlalchemy = "^2.0"
dependency-injector = "^4.48.1"
aiohttp = "^3.12.15"
watchtower = "^3.4.0"
psycopg2-binary = "^2.9.10"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
boto3 = "^1.40.15"
botocore = "^1.40.15"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
faker = "^37.5.3"
factory-boy = "^3.3.3"

[tool.pytest.ini_options]
testpaths = "tests"
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=html
  --cov-report=xml
  --cov-fail-under=100
  --numprocesses=1
  --asyncio-mode=auto
'''

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

